Optimized filter implementation
Small fix

More work, start writing filter -> selection vector

Things compiling again finally

BinaryBitBlockCounter tests passing

Consolidate take/filter tests in same module, fix GetTakeIndices / GetFilterOutputSize unit tests and implementations

Finish filter implementation, tests passing again

Clean up includes

Tweak benchmark parameters

Some string streamlining

Python fixes

Python test fixes. Add fast path for low-selectivity filters

Low selectivity path for non-primitive filtering

VisitFilter is not a dependent template

Implement some obvious non-null filter optimizations

Fix typo
