+++ cpp/src/arrow/CMakeLists.txt
@@ -356,10 +356,10 @@ if(ARROW_COMPUTE)
               compute/kernels/scalar_set_lookup.cc
               compute/kernels/scalar_string.cc
               compute/kernels/scalar_validity.cc
-              compute/kernels/vector_filter.cc
+              compute/kernels/util_internal.cc
               compute/kernels/vector_hash.cc
-              compute/kernels/vector_sort.cc
-              compute/kernels/vector_take.cc)
+              compute/kernels/vector_selection.cc
+              compute/kernels/vector_sort.cc)
 endif()
 
 if(ARROW_FILESYSTEM)
