+++ cpp/src/arrow/compute/kernels/CMakeLists.txt
@@ -38,9 +38,8 @@ add_arrow_benchmark(scalar_string_benchmark PREFIX "arrow-compute")
 
 add_arrow_compute_test(vector_test
                        SOURCES
-                       vector_filter_test.cc
                        vector_hash_test.cc
-                       vector_take_test.cc
+                       vector_selection_test.cc
                        vector_sort_test.cc
                        test_util.cc)
